ai/apps/DG_collect_dataset/core/requirements.txt
# --- Core AI Stack ---
torch>=2.4.0
torchvision>=0.19.0
transformers>=4.46.0
accelerate>=0.26.0
bitsandbytes>=0.42.0
qwen-vl-utils>=0.0.8

# --- Image Processing ---
insightface>=0.7.3
onnxruntime-gpu>=1.17.0
opencv-python>=4.8.0
Pillow>=10.0.0
simple-lama-inpainting>=0.1.0

# --- Automation & Utilities ---
numpy>=1.24.0
pandas>=2.0.0
tqdm>=4.66.0
playwright>=1.40.0
beautifulsoup4>=4.12.0
requests>=2.31.0
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.0.0